requirements.txt
numpy==1.24.4
pandas==2.0.3
Pillow==10.0.1
poolformer==0.0.4
scikit_learn==1.3.1
scikit-image
timm
torch==2.1.0
torchvision==0.16.0
wandb==0.15.12
